hw02_unpack_string/unpack_test.go
package hw02_unpack_string //nolint:golint,stylecheck

import (
	"testing"

	"github.com/stretchr/testify/require"
)

type test struct {
	input    string
	expected string
	err      error
}

func TestUnpack(t *testing.T) {
	for _, tst := range [...]test{
		{
			input:    "a4bc2d5e",
			expected: "aaaabccddddde",
		},
		{
			input:    "abcd",
			expected: "abcd",
		},
		{
			input:    "45",
			expected: "",
			err:      ErrInvalidString,
		},
		{
			input:    "aaa10b",
			expected: "",
			err:      ErrInvalidString,
		},
		{
			input:    "",
			expected: "",
		},
		{
			input:    "a2o3f0",
			expected: "",
			err:      ErrInvalidString,
		},
		{
			input:    "a2c000",
			expected: "",
			err:      ErrInvalidString,
		},
		{
			input: "!@#!@%@$!@#",
			expected: "",
			err: ErrInvalidString,
		},
		{
			input: "a100000000000000000000000000",
			expected: "",
			err: ErrInvalidString,
		},
		{
			input: "♬1♪2",
			expected: "",
			err: ErrInvalidString,
		},
		{
			input: "д2р3ф2",
			expected: "ддрррфф",
		},
	} {
		t.Run(tst.input, func(t *testing.T) {
			result, err := Unpack(tst.input)
			require.Equal(t, tst.err, err)
			require.Equal(t, tst.expected, result)
		})
	}
}

func TestUnpackWithEscape(t *testing.T) {
	t.Skip() // Remove if task with asterisk completed

	for _, tst := range [...]test{
		{
			input:    `qwe\4\5`,
			expected: `qwe45`,
		},
		{
			input:    `qwe\45`,
			expected: `qwe44444`,
		},
		{
			input:    `qwe\\5`,
			expected: `qwe\\\\\`,
		},
		{
			input:    `qwe\\\3`,
			expected: `qwe\3`,
		},
	} {
		t.Run(tst.input, func(t *testing.T) {
			result, err := Unpack(tst.input)
			require.Equal(t, tst.err, err)
			require.Equal(t, tst.expected, result)
		})
	}
}
